diabetic-retinopathy-detection/requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning>=2.0.0

# Model architectures
timm>=0.9.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
Pillow>=9.5.0
albumentations>=1.3.0
opencv-python-headless>=4.8.0

# Metrics and evaluation
scikit-learn>=1.3.0

# Configuration
PyYAML>=6.0
omegaconf>=2.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Logging and experiment tracking
tensorboard>=2.13.0
rich>=13.0.0

# Optional: Weights & Biases integration
# wandb>=0.15.0

# Optional: Jupyter notebook support
# jupyter>=1.0.0
# ipywidgets>=8.0.0

# Development
tqdm>=4.65.0
